requirements.txt
alembic==1.4.3
click==7.1.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Markdown==0.3
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
Pillow==7.2.0
pkg-resources==0.0.0
PyMySQL==0.10.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-slugify==4.0.1
six==1.15.0
SQLAlchemy==1.3.19
text-unidecode==1.3
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.2.0
